+++ pydash/__init__.py
@@ -163,6 +163,7 @@ from .functions import (
     flow_right,
     iterated,
     juxtapose,
+    mod_args,
     negate,
     once,
     partial,