+++ tests/test_numerical.py
@@ -40,6 +40,26 @@ def test_average_aliases(case):
 
 
 @parametrize('case,expected', [
+    ((4.006,), 5),
+    ((6.004, 2), 6.01),
+    ((6040, -2), 6100),
+    (([4.006, 6.004], 2), [4.01, 6.01]),
+])
+def test_ceil(case, expected):
+    assert _.ceil(*case) == expected
+
+
+@parametrize('case,expected', [
+    ((4.006,), 4),
+    ((0.046, 2), 0.04),
+    ((4060, -2), 4000),
+    (([4.006, 0.046], 2), [4.0, 0.04]),
+])
+def test_floor(case, expected):
+    assert _.floor(*case) == expected
+
+
+@parametrize('case,expected', [
     (([0, 0, 0, 0, 5],), 0),
     (([0, 0, 1, 2, 5],), 1),
     (([0, 0, 1, 2],), 0.5),
@@ -97,7 +117,7 @@ def test_round_(case, expected):
 @parametrize('case', [
     _.curve
 ])
-def test_round__aliases(case):
+def test_round_aliases(case):
     assert _.round_ is case