+++ tests/test_objects.py
@@ -120,13 +120,6 @@ def test_defaults_deep(case, expected):
     assert _.defaults_deep(*case) == expected
 
 
-@parametrize('case', [
-    _.deep_defaults
-])
-def test_defaults_deep_aliases(case):
-    assert _.defaults_deep is case
-
-
 @parametrize('case,expected', [
     ([1, 2, 3], {0: 1, 1: 2, 2: 3}),
     ({0: 1, 1: 2, 2: 3}, {0: 1, 1: 2, 2: 3}),