+++ CHANGES.rst
@@ -4,6 +4,7 @@ Changelog
 =========
 
 
+- Add ``mod_args``.
 - Add ``unzip_with``.
 - Add ``zip_with``.