+++ CHANGES.rst
@@ -6,6 +6,8 @@ Changelog
 
 - Add ``ceil``.
 - Add ``floor``.
+- Add ``method``.
+- Add ``method_of``.
 - Add ``mod_args``.
 - Add ``unzip_with``.
 - Add ``zip_with``.