+++ CHANGES.rst
@@ -17,10 +17,12 @@ Changelog
 - Add ``method``.
 - Add ``method_of``.
 - Add ``mod_args``.
+- Add ``set_``.
 - Add ``unzip_with``.
 - Add ``zip_with``.
 - Make ``add`` support adding two numbers if passed in positionally.
 - Make ``get`` main definition and ``get_path`` its alias.
+- Make ``set_`` main definition and ``deep_set`` its alias.
 
 
 v3.2.2 (2015-04-29)